+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-# 0.1.0
+# 0.2.1
+
+- Added `maybeCompact` -> Optional in `add,addAll,update,delete`
+- Added `isDBInitialized` -> init Usage
+- Added `addAll` -> `Multi Add Record`
+
+## 0.1.0
 
 - Initial version.